Move ultralcd related stuff to ultralcd.cpp/h
